requirements.txt
redis>=2.10.5
redis-shard>=0.3.3
falcon>=1.0.0
requests>=2.9.1
Theano>=0.8.2
Keras>=1.0.4
DBUtils>=1.1
thriftpy>=0.3.8
grequests>=0.3.0
gsocketpool>=0.1.5
gunicorn>=19.6.0
gevent-etcd>=2.3
gevent-websocket>=0.9.5
geventconnpool>=0.2.1
geventcron>=1.0
geventhttpclient>=1.2.0
ansible>=1.9.4
algorithms>=1.0
pyangbind>=0.5.4
ryu>=4.3
schedule>=0.3.2
xmltodict>=0.10.2
tushare>=0.5.0
fabric>=1.11.1
fabric-gunicorn>=0.1
kafka-python>=1.2.1
lggr>=0.2.0
msgpack-python>=0.3.0
eventlet>=0.15
autojenkins>=1.1.0
artifactory>=0.1.17
peewee>=2.8.1
beautifulsoup4>=4.4.1
html5lib>=0.9999999
socialoauth>=0.3.3
weibo>=0.2.2
wechatpy>=1.2.12
pyTOP>=0.1.2
alipay>=0.7.3
paypal>=1.2.5
upyun>=2.3.2
sevencow>=0.2.0
python-zipkin>=0.1.0
python-aliyun>=1.1.0
pywinauto>=0.5.4
saws>=0.4.1
prompt-toolkit>=1.0.3
aws-shell>=0.1.1
grpcio>=0.14.0
retrying>=1.3.3
pymemcache>=1.3.5
thrift-tools>=0.0.3
cachetools>=1.1.6
kafka-utils>=0.2.0
yapf>=0.10.0
pymongo>=3.2.2
mongoengine>=0.10.6
python-dateutil>=2.5.3
pillow>=3.2.0
netaddr>=0.7.18
python-etcd>=e0.4.3
pyleus>=0.3.0
petrel>=1.0.1.0.3
streamparse>=3.0.0
pysparkling>=0.3.22
sparkit-learn 0.2.6
psutil>=4.3.0
#you-get>=0.4.455
youtube_dl>=2016.7.7
jieba>=0.38
boto>=2.41.0
pika>=0.10.0
luigi>=2.1.1
IPy>=0.83
xlwt>=1.1.2
xlutils>=2.0.0
xlrd>=1.0.0
XlsxWriter>=0.9.3
python-docx>=0.8.6